Agrega el componente MovimientoItem para mostrar detalles de movimientos. Se implementa la lógica para determinar el icono según la forma de pago y se mejora la presentación de la información en el modal.
